+++ lib/reducers_test.ts
@@ -100,3 +100,9 @@ Deno.test("findIndex", () => {
     6,
   );
 });
+
+Deno.test("includes", () => {
+  const numbers = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
+  assertEquals(reducers.includes(numbers, 69), false);
+  assertEquals(reducers.includes(numbers, 5), true);
+});